GraphQL Refactor Step 4 - Refactor Transform Invocation into New Top-Level Package (#1550)
* chore: migrate root transformer package into new npm package

* fix: update imports, and add back schema-generator to category deps

* chore: use strict ts in new package
