+++ packages/amplify-category-api/package.json
@@ -30,16 +30,8 @@
   },
   "dependencies": {
     "@aws-amplify/graphql-auth-transformer": "2.1.6",
-    "@aws-amplify/graphql-default-value-transformer": "1.1.6",
-    "@aws-amplify/graphql-function-transformer": "1.2.4",
-    "@aws-amplify/graphql-http-transformer": "1.2.4",
-    "@aws-amplify/graphql-index-transformer": "1.2.2",
-    "@aws-amplify/graphql-maps-to-transformer": "2.1.6",
-    "@aws-amplify/graphql-model-transformer": "1.3.2",
-    "@aws-amplify/graphql-predictions-transformer": "1.2.4",
-    "@aws-amplify/graphql-relational-transformer": "1.2.2",
     "@aws-amplify/graphql-schema-generator": "0.1.2",
-    "@aws-amplify/graphql-searchable-transformer": "1.2.4",
+    "@aws-amplify/graphql-transformer": "0.0.1",
     "@aws-amplify/graphql-transformer-core": "1.3.2",
     "@aws-amplify/graphql-transformer-interfaces": "2.2.1",
     "@aws-amplify/graphql-transformer-migrator": "2.1.6",
