--- packages/amplify-category-api/src/amplify-graphql-transform/index.ts
@@ -1,1 +1,0 @@
-export { constructTransformerChain, constructTransform } from './graphql-transformer-v2';
