+++ packages/amplify-category-api/src/graphql-transformer/directive-definitions.ts
@@ -4,7 +4,7 @@ import { print } from 'graphql';
 import { TransformerPluginProvider } from '@aws-amplify/graphql-transformer-interfaces';
 import { getTransformerFactoryV1 } from './transformer-factory';
 import { getTransformerVersion } from './transformer-version';
-import { constructTransformerChain } from '../amplify-graphql-transform';
+import { constructTransformerChain } from '@aws-amplify/graphql-transformer';
 import { loadCustomTransformersV2 } from './transformer-options-v2';
 
 /**
