+++ packages/amplify-category-api/src/graphql-transformer/override.ts
@@ -4,7 +4,7 @@ import * as vm from "vm2";
 import * as path from "path";
 import _ from "lodash";
 import { pathManager, stateManager } from "@aws-amplify/amplify-cli-core";
-import { StackManager } from "@aws-amplify/graphql-transformer-core";
+import { StackManager } from '@aws-amplify/graphql-transformer-core';
 import { AmplifyApiGraphQlResourceStackTemplate } from "@aws-amplify/graphql-transformer-interfaces";
 import { ConstructResourceMeta } from "./types/types";
 import { convertToAppsyncResourceObj, getStackMeta } from "./types/utils";
