+++ packages/amplify-category-api/src/graphql-transformer/transform-graphql-schema-v2.ts
@@ -24,7 +24,7 @@ import { generateTransformerOptions } from './transformer-options-v2';
 import { TransformerProjectOptions } from './transformer-options-types';
 import { getExistingConnectionSecretNames, getSecretsKey } from '../provider-utils/awscloudformation/utils/rds-secrets/database-secrets';
 import { getAppSyncAPIName } from '../provider-utils/awscloudformation/utils/amplify-meta-utils';
-import { executeTransform } from '../amplify-graphql-transform/graphql-transformer-v2';
+import { executeTransform } from '@aws-amplify/graphql-transformer';
 
 const PARAMETERS_FILENAME = 'parameters.json';
 const SCHEMA_FILENAME = 'schema.graphql';
