+++ packages/amplify-graphql-transformer/.npmignore
@@ -1,0 +1,5 @@
+**/__mocks__/**
+**/__tests__/**
+./src
+tsconfig.json
+tsconfig.tsbuildinfo
