+++ packages/amplify-graphql-transformer/src/index.ts
@@ -1,0 +1,11 @@
+export {
+  constructTransformerChain,
+  constructTransform,
+  executeTransform,
+} from './graphql-transformer';
+
+export type {
+  TransformerFactoryArgs,
+  TransformConfig,
+  ExecuteTransformConfig,
+} from './graphql-transformer';
