+++ packages/amplify-graphql-transformer/tsconfig.json
@@ -1,0 +1,14 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "lib",
+    "rootDir": "src",
+    "strict": true,
+    "allowJs": false
+  },
+  "exclude": [
+    "coverage",
+    "lib",
+    "src/__tests__"
+  ]
+}
